Initial successful test.
Able to authenticate to an ArcGIS Server to both the 'consumption' endpoints and the back-end administration endpoints.

Future TODO's:

Acqure token once, then track expiration and re-acquire as needed for long-running tasks/processes...

Ensure token is protected.  Would like to get the token in the body of a post request, however challenges forced placing it as a 'parameter'.

Issues with re-direction.  The underlying HTTP Method override (to POST) and storing token in the body is lost.